+++ R/use_extendr.R
@@ -120,6 +120,20 @@ use_extendr <- function(
   )
 
   use_rextendr_template(
+    "cleanup",
+    save_as = file.path("src", "cleanup"),
+    quiet = quiet,
+    overwrite = overwrite
+  )
+
+  use_rextendr_template(
+    "cleanup.win",
+    save_as = file.path("src", "cleanup.win"),
+    quiet = quiet,
+    overwrite = overwrite
+  )
+
+  use_rextendr_template(
     "_gitignore",
     save_as = file.path("src", ".gitignore"),
     quiet = quiet,